making new db functions and db on localhost first before putting on supabase
